pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Piranha887 <ilyanikitin889@gmail.com>"]
readme = "README.md"
repository = "https://github.com/piranha887/python-project-50"
packages = [
	{ include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"


[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
